Refactor: 잘못된 패키지명 변경 (db -> entity)
