+++ src/main/java/com/cocos/cocos/api/review/service/ReviewService.java
@@ -20,7 +20,11 @@ import com.cocos.cocos.api.review.dto.response.*;
 import com.cocos.cocos.db.member.repository.MemberRepository;
 import com.cocos.cocos.db.pet.entity.Pet;
 import com.cocos.cocos.db.pet.repository.PetRepository;
-import com.cocos.cocos.db.review.db.*;
+import com.cocos.cocos.db.review.entity.Review;
+import com.cocos.cocos.db.review.entity.ReviewImage;
+import com.cocos.cocos.db.review.entity.ReviewSummary;
+import com.cocos.cocos.db.review.entity.ReviewSummaryOption;
+import com.cocos.cocos.db.review.entity.ReviewSymptom;
 import com.cocos.cocos.db.review.repository.*;
 import com.cocos.cocos.enums.message.FailMessage;
 import com.cocos.cocos.db.symptom.entity.Symptom;
