+++ src/main/java/com/cocos/cocos/db/review/repository/ReviewImageRepository.java
@@ -1,6 +1,6 @@
 package com.cocos.cocos.db.review.repository;
 
-import com.cocos.cocos.db.review.db.ReviewImage;
+import com.cocos.cocos.db.review.entity.ReviewImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
