+++ src/main/java/com/cocos/cocos/db/review/repository/ReviewSummaryOptionRepository.java
@@ -1,6 +1,6 @@
 package com.cocos.cocos.db.review.repository;
 
-import com.cocos.cocos.db.review.db.ReviewSummaryOption;
+import com.cocos.cocos.db.review.entity.ReviewSummaryOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
