+++ src/main/java/com/cocos/cocos/db/review/repository/ReviewSummaryRepository.java
@@ -1,6 +1,6 @@
 package com.cocos.cocos.db.review.repository;
 
-import com.cocos.cocos.db.review.db.ReviewSummary;
+import com.cocos.cocos.db.review.entity.ReviewSummary;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
