+++ src/main/java/com/cocos/cocos/db/review/repository/ReviewSymptomRepository.java
@@ -1,6 +1,6 @@
 package com.cocos.cocos.db.review.repository;
 
-import com.cocos.cocos.db.review.db.ReviewSymptom;
+import com.cocos.cocos.db.review.entity.ReviewSymptom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
